requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
cffi==1.16.0
cryptography==41.0.4
decorator==5.1.1
Django==3.2.18
django-admin-honeypot==1.1.0
django-user-agents==0.4.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
executing==2.0.0
ipdb==0.13.13
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.11.2
typing_extensions==4.8.0
ua-parser==0.18.0
user-agents==2.2.0
wcwidth==0.2.8
